move legacy_caching
