+++ lib/streamlit/delta_generator.py
@@ -34,8 +34,7 @@ from typing_extensions import Final, Literal
 
 import streamlit as st
 from streamlit import cursor
-from streamlit.runtime import caching
-from streamlit import legacy_caching
+from streamlit.runtime import caching, legacy_caching
 from streamlit import type_util
 from streamlit import util
 from streamlit.cursor import Cursor
