+++ lib/tests/streamlit/runtime/app_session_test.py
@@ -117,7 +117,7 @@ class AppSessionTest(unittest.TestCase):
         session._handle_clear_cache_request()
         self.assertTrue("foo" not in session._session_state)
 
-    @patch("streamlit.legacy_caching.clear_cache")
+    @patch("streamlit.runtime.legacy_caching.clear_cache")
     @patch("streamlit.runtime.caching.memo.clear")
     @patch("streamlit.runtime.caching.singleton.clear")
     def test_clear_cache_all_caches(
