+++ lib/tests/streamlit/runtime/scriptrunner/script_runner_test.py
@@ -26,7 +26,7 @@ from tornado.testing import AsyncTestCase
 
 from streamlit import source_util
 from streamlit.elements.exception import _GENERIC_UNCAUGHT_EXCEPTION_TEXT
-from streamlit.legacy_caching import caching
+from streamlit.runtime.legacy_caching import caching
 from streamlit.proto.ClientState_pb2 import ClientState
 from streamlit.proto.Delta_pb2 import Delta
 from streamlit.proto.Element_pb2 import Element
